remove numpy from build requirements
The issue is that pip builds the package in an isolated environment
where it systematically downloads all those requirements,
so the version of numpy used to build the package may mismatch the
version of the installed numpy.

To prevent this, we could do ``pip install . -v --no-build-isolation``
but then it requires that all build dependencies (cmake, scikit-build)
be installed explicitly by the user.

By removing numpy here, we only require that the user install numpy
before, which is more reasonable IMO.

See
https://github.com/pypa/pip/issues/6193
https://github.com/pypa/pip/issues/6144